Move signals into own package
