+++ signals/signals.go
@@ -1,4 +1,4 @@
-package main
+package signals
 
 import (
 	"fmt"
@@ -18,9 +18,9 @@ func init() {
 	ValidSignals = valid
 }
 
-// ParseSignal parses the given string as a signal. If the signal is not found,
+// Parse parses the given string as a signal. If the signal is not found,
 // an error is returned.
-func ParseSignal(s string) (os.Signal, error) {
+func Parse(s string) (os.Signal, error) {
 	sig, ok := SignalLookup[strings.ToUpper(s)]
 	if !ok {
 		return nil, fmt.Errorf("invalid signal %q - valid signals are %q",
